requirements.txt
asgiref==3.8.1
Django==4.2.19
gunicorn==23.0.0
numpy==2.0.2
pillow==11.1.0
sqlparse==0.5.3
torch==2.6.0
tzdata==2025.1